requirements.txt
psycopg2
streamlit
pandas
pyrebase4
